+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- `setStore` now accepts default options for each option type on the
+  `DelegatedStoreOptions`: `setItemOptions`, `listItemsOptions`,
+  `getItemsOptions`, and `batchOptions`. These are merged with (and overridden
+  by) any per-call options. `DelegatedStoreConfig` now extends
+  `DelegatedStoreOptions`.
+
 ## [0.13.0] - 2026-06-01
 
 ### BREAKING Changes